Factor out usePermalinkTargetRoom
